Rename lib file to appease rubocop
