+++ bin/deploy-entitlements
@@ -2,7 +2,7 @@
 
 ENV["BUNDLE_GEMFILE"] = File.expand_path("../Gemfile", File.dirname(__FILE__))
 require "bundler/setup"
-require "ruby-version-check"
+require "ruby_version_check"
 
 # Note that contracts.ruby has two specific ruby-version specific libraries, which we have vendored into lib/
 if RubyVersionCheck.ruby_version2?
