+++ spec/unit/spec_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "ruby-version-check"
+require "ruby_version_check"
 require "base64"
 
 # Note that contracts.ruby has two specific ruby-version specific libraries, which we have vendored into lib/
